hooks extracted into separate files, top-5 rated animes added on Main
