+++ src/components/Card/Card.module.css
@@ -56,7 +56,7 @@
     text-align: center;
 }
 
-.btn {
+.view_btn {
     margin-bottom: 10px;
     width: 165px;
 }
